fix(analytics): gate engine_changed on real engine change; f64 duration; props:{} on correction_learned

+++ src-tauri/src/recording.rs
@@ -369,7 +369,7 @@ pub(crate) async fn stop_and_transcribe_internal<R: Runtime>(app: AppHandle<R>) 
     // Notify the widget FIRST so it starts its close countdown.
     app.emit("transcription-done", &text).ok();
     let engine = if is_apple { "apple" } else { "whisper" };
-    let duration_ms = (duration_secs * 1000.0) as i64;
+    let duration_ms = (duration_secs as f64 * 1000.0) as i64;
     let props = crate::analytics::transcription_props(engine, &model_name, &language, "dictation", &text, duration_ms);
     app.emit("analytics-event", serde_json::json!({ "event": "transcription_completed", "props": props })).ok();
     // 300ms: time for the previously-active app to regain keyboard focus before Cmd+V.
@@ -541,7 +541,7 @@ pub async fn transcribe_file<R: Runtime>(app: AppHandle<R>, path: String) -> Res
     // 7. Done — no paste (file transcription is not a dictation).
     app.emit("file-transcribe-done", &text).ok();
     let engine = if is_apple { "apple" } else { "whisper" };
-    let duration_ms = (samples.len() as f32 / 16000.0 * 1000.0) as i64;
+    let duration_ms = (samples.len() as f64 / 16000.0 * 1000.0) as i64;
     let props = crate::analytics::transcription_props(engine, &model_name, &language, "file", &text, duration_ms);
     app.emit("analytics-event", serde_json::json!({ "event": "transcription_completed", "props": props })).ok();
     Ok(text)
